feat: made elements helper and added error handling for code formatting
