+++ src/actions/useTurnInto.actions.tsx
@@ -1,6 +1,6 @@
 import useRnbw from "@_services/useRnbw";
 import { Range, editor } from "monaco-editor";
-import { useElementHelper } from "@_services/useElementHelper";
+import { PrettyCode } from "@_services/useElementsHelper";
 
 // helperModel added to update the code in the codeViewInstanceModel
 // once when the action is executed, this improves the History Management
@@ -8,7 +8,7 @@ const helperModel = editor.createModel("", "html");
 
 export default function useTurnInto() {
   const rnbw = useRnbw();
-  const { PrettyCode } = useElementHelper();
+
   const selectedElements = rnbw.elements.getSelectedElements();
 
   async function turnInto(tagName: string) {
